fix : fixed the logic of comparator function that's used to sort notifications(desc order) on the basis of their timestamps.
The comparator function that is called in DatabaseHelper.kt to sort the notifications in Descending order of their timestamp
is incorrect and doesn't provide the right comparator function logic to sort objects in descending order. According to online java/kotlin
documentation a comparator function should provide return values for all three cases.

1. updated the comparator function to provide return values for all three conditions(a > b ,a < b and a == b), where a and b are timestamps
of any two notifications that are being compared.