fix #1499 : making on-screen/soft keyboard dismiss on background tap
To enhance the user flow, added the ability to dismiss the soft keyboard or on-screen keyboard by
tapping on the background or anywhere other than an EditText.

1. dismissSoftKeyboardOnBkgTap() takes View/ViewGroup as an input and then individually checks
all its child Views/ViewGroups to find a View that doesn't include EditText.

2. when such a view is found and a tap is made on such a view, it's passed to hideSoftKeyboard() Utils function
that by using Input Method Manager hides the soft keyboard.

3. tested on pixel 2(api 33), the fix works smoothly.