ci: Generate Firebase version and deploy to Firebase App Distribution
This commit introduces a new lane `generateFirebaseVersion` that generates a new version code for Firebase App Distribution by incrementing the latest build version. It also updates the `deploy_on_firebase` lane to use the demo release APK and adds `generateFirebaseVersion` to the main `build` lane.

Additionally, the existing `generateVersion` lane is renamed to `generatePlayStoreVersion` to better reflect its purpose, which is generating a version code for the Play Store.